test: create new tests for each version
Tests are not fully functional yet, but build the base to interate on
